refactor: replace file mode with test mode in edit (#2952)
