+++ vscode/src/commands/execute/explain.ts
