+++ vscode/src/commands/execute/smell.ts
