+++ vscode/src/edit/types.ts
@@ -8,6 +8,6 @@ export type EditIntent = 'add' | 'edit' | 'fix' | 'doc' | 'new'
  * The edit modes that can be used when applying an edit.
  * - 'edit': Modify selected code in place.
  * - 'insert': Insert new code above the selected code.
- * - 'file': Create a new file and insert code there..
+ * - 'test': Create a new test file and insert tests there. Used for unit-test command only.
  */
-export type EditMode = 'edit' | 'insert' | 'file'
+export type EditMode = 'edit' | 'insert' | 'test'
